backend-app/requirements.txt
annotated-types==0.6.0
anyio==4.3.0
bcrypt==4.1.2
certifi==2024.6.2
click==8.1.7
coverage==7.5.4
ecdsa==0.18.0
et-xmlfile==1.1.0
fastapi==0.110.0
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.6
iniconfig==2.0.0
numpy==1.26.4
openpyxl==3.1.2
packaging==24.0
pandas==2.2.1
passlib==1.7.4
pluggy==1.4.0
psycopg2-binary==2.9.9
pyasn1==0.5.1
pydantic==2.6.4
pydantic_core==2.16.3
pytest==8.1.1
pytest-cov==5.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.9
pytz==2024.1
rsa==4.9
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.29
starlette==0.36.3
typing_extensions==4.10.0
tzdata==2024.1
uvicorn==0.28.0
